racket--calcualate-indent: Don't return a list
Decide that racket--calcualate-indent may only return an integer or
nil.

By contrast, calculate-lisp-indent may also return a list of two
elements. From its doc string:

> The value can also be a list of the form (COLUMN
> CONTAINING-SEXP-START). This means that following lines at the same
> level of indentation should not necessarily be indented the same as
> this line. Then COLUMN is the column to indent to, and
> CONTAINING-SEXP-START is the buffer position of the start of the
> containing expression."

AFACIT this is used only by indent-pp-sexp, which in lisp-mode is
bound to C-M-q.

However in racket-mode C-M-q is bound to prog-indent-sexp -- which
calls racket-indent-line (via indent-region and indent-line-function).
And we don't actually use the list's second element.