Simplify racket--calculate-indent
Delete whole chunk of code designed to move point before calling
racket-indent-function -- which doesn't use the position of point.
This functionality of calculate-lisp-indent may be necessary for it to
preserve because lisp-indent-functions may depend on it. But that's
N/A for us.

The result can be refactored even more, but doing this interim commit
as a checkpoint (plan to squash all this eventually).